move NodeSearch to searches directory
