+++ src/components/searches/index.ts
@@ -1,1 +1,2 @@
 export * from './StorySearch';
+export * from './NodeSearch';
